added: nice looking two column layout & updated the views to reflect it.  (it needs some navigational work and we probably should add a sponsor listing to the side bar...)
changed:  some of the views had untyped view data.  updated to use ViewPage<T> so the viewdata is stronly typed (which avoids the ugly script runat=server in the view to get a typed variable.) -- in some cases where more than one object is being used by the view, I created a XX_ViewData class that lives in the website/models folder.  (should this be in Domain.Presentation?)